refactor: small refactor
