+++ src/ts/main-page/mainFilters.ts
@@ -1,5 +1,5 @@
-import { dataProducts } from "./dataProducts";
-import { addQuery } from "./Router";
+import { dataProducts } from "../dataProducts";
+import { addQuery } from "../Router";
 interface CustomHtmlDiv extends HTMLDivElement  {
     type?: string;
 }
